Match ids of index and evaluate htmls
Since we're using the same js for both the index and evaluate
files, we have to make sure the ids of the divs that the js
for both pages interacts with exist in both html files. We can
throw them in a hidden div to make sure they don't actually
show up.